+++ kernel/raster/fb.c
@@ -516,7 +516,7 @@ void fb_init_indexed(
     int width,
     int height,
     int pitch,
-    int8_t bpp
+    int bpp
 ) {
     s_colorinfo.indexed.palette = palette;
     s_colorinfo.indexed.colorcount = colorcount;
